Add getCoords function to mariadb data source
Implemented a new getCoords function in the mariadb data source and connected it to the GraphQL schema. This function retrieves the 'x' and 'y' coordinates of an event, thus providing a way to fetch spatial data from the Database. Modified the Event type in the GraphQL schema to contain a Point object that includes the 'x' and 'y' coordinates.